Refactor saveAllFiles
